+++ modules/todolist/widgets/iconbutton.h
@@ -7,10 +7,10 @@ class IconButton : public QPushButton
 {
 
 public:
-    IconButton(std::function<void()> onClic, QString name, QWidget *p, size_t size = 30);
+    IconButton(std::function<void()> onClic, QString name, QWidget *p, int size = 30);
 
 private:
-    size_t size;
+    int size;
 
     virtual void enterEvent(QEnterEvent *);
     virtual void leaveEvent(QEvent *);
